Refactored file inclusion in few .h and .cpp files
