+++ Graphics/Cubemap.h
@@ -1,23 +1,5 @@
 #pragma once
-
-namespace std
-{
-	template <class _Ty>
-	class allocator;
-	template <class _Ty, class _Alloc = allocator<_Ty>>
-	class vector;
-
-	template <class _Elem>
-	struct char_traits;
-	template <class _Elem, class _Traits = char_traits<_Elem>, class _Alloc = allocator<_Elem>>
-	class basic_string;
-
-	using string = basic_string<char, char_traits<char>, allocator<char>>;
-}
-struct Vertex;
-class VertexArrayObject;
-class ShaderProgram;
-typedef unsigned int GLuint;
+#include "BaseHeader.h"
 
 class Cubemap
 {
