+++ Graphics/ShaderProgram.h
@@ -1,7 +1,5 @@
 #pragma once
-#include <glad/glad.h>
-#include <glm/glm.hpp>
-#include <string>
+#include "BaseHeader.h"
 
 class ShaderProgram
 {
