+++ Graphics/Texture.h
@@ -1,5 +1,5 @@
 #pragma once
-#include <glad/glad.h>
+#include "BaseHeader.h"
 #include <string>
 
 class Texture
