+++ Graphics/Window.h
@@ -1,6 +1,5 @@
 #pragma once
-#include <glad/glad.h>
-#include <glm/glm.hpp>
+#include "BaseHeader.h"
 
 class GLFWwindow;
 class Window
